Initial implementation of RTA scraper.
The RTA site embeds the meeting schedule using an iframe, so the tests
are currently just using the source of the iframe to test item
extraction. In order to obtain more thorough coverage, we'll need to
consider using a different testing strategy that can stub out Requests
creates during the test run.

Also note that the meeting time is currently hardcoded, which is not
ideal but I think OK for our immediate purposes.

Scrapy supports Downloader middleware, so either the HttpCacheMiddleware
or a simple custom one will probably do the job.