Retry attempts to fetch a URL.
Our intermittent test failures are the result of the citybureau.org
server returning a 429 response when a test attempts to fetch some
HTML. In an effort to make our tests solidly green and remove the
intermittent failures, this commit adds retrying functionality to the
_gen_html function. We retry up to 5 times, with an exponentially increasing
backoff (2, 4, 8, 16, 32 seconds).

Please note that this is a temporary fix, as ultimately we should not be hitting
the network to run these tests. I can think of two solutions:

1. Have two test suites, one for local unit tests and one that tests
   remote systems.
2. Use a library to mock out these requests so we don't rely on the
   network.

In the short term, however, I think this fix will improve the experience
for new developers.